moved clients functionality to separate clients django app
