+++ clients/templates/clients/client_form.html
@@ -23,7 +23,7 @@
     <h1>{% translate 'Client form' %}</h1>
     <ul>
       <li><a href="{% url 'main' lang=lang %}">{% translate 'Main' %}</a></li>
-      <li><a href="{% url 'handbooks:all_client_list' lang=lang %}">{% translate 'Clients' %}</a></li>
+      <li><a href="{% url 'clients:all_client_list' lang=lang %}">{% translate 'Clients' %}</a></li>
       <li>/ </li>
       <li>{% translate 'Client form' %}</li>
     </ul>
