+++ clients/templates/clients/client_list.html
@@ -48,7 +48,7 @@
     <ul class="nav nav-tabs" id="myTab" role="tablist">
         <li class="nav-item" role="presentation">
             <a class="nav-link active"
-               href="{% url 'handbooks:all_client_list' lang=lang %}">
+               href="{% url 'clients:all_client_list' lang=lang %}">
                 {% translate 'Clients' %}
             </a>
         </li>
@@ -73,37 +73,37 @@
         <ul class="nav nav-tabs" id="myTab1" role="tablist">
             <li class="nav-item" role="presentation">
                 <a class="nav-link {% if url_name == 'all_client_list' %}active{% endif %}"
-                   href="{% url 'handbooks:all_client_list' lang=lang %}">
+                   href="{% url 'clients:all_client_list' lang=lang %}">
                     {% translate 'All' %}
                 </a>
             </li>
             <li class="nav-item" role="presentation">
                 <a class="nav-link {% if url_name == 'new_client_list' %}active{% endif %}"
-                   href="{% url 'handbooks:new_client_list' lang=lang %}">
+                   href="{% url 'clients:new_client_list' lang=lang %}">
                     {% translate 'New' %}
                 </a>
             </li>
             <li class="nav-item" role="presentation">
                 <a class="nav-link {% if url_name == 'in_selection_client_list' %}active{% endif %}"
-                   href="{% url 'handbooks:in_selection_client_list' lang=lang %}">
+                   href="{% url 'clients:in_selection_client_list' lang=lang %}">
                     {% translate 'In selection' %}
                 </a>
             </li>
             <li class="nav-item" role="presentation">
                 <a class="nav-link {% if url_name == 'with_show_client_list' %}active{% endif %}"
-                   href="{% url 'handbooks:with_show_client_list' lang=lang %}">
+                   href="{% url 'clients:with_show_client_list' lang=lang %}">
                     {% translate 'With a show' %}
                 </a>
             </li>
             <li class="nav-item" role="presentation">
                 <a class="nav-link {% if url_name == 'decided_client_list' %}active{% endif %}"
-                   href="{% url 'handbooks:decided_client_list' lang=lang %}">
+                   href="{% url 'clients:decided_client_list' lang=lang %}">
                     {% translate 'Decided' %}
                 </a>
             </li>
             <li class="nav-item" role="presentation">
                 <a class="nav-link {% if url_name == 'deferred_demand_client_list' %}active{% endif %}"
-                   href="{% url 'handbooks:deferred_demand_client_list' lang=lang %}">
+                   href="{% url 'clients:deferred_demand_client_list' lang=lang %}">
                     {% translate 'Deferred demand' %}
                 </a>
             </li>
@@ -115,7 +115,7 @@
 
                 {% if can_add_client %}
                     <a class="btn btn-primary btn-sm btn-warning element-fullwidth"
-                       href="{% url 'handbooks:client_create' lang=lang %}"
+                       href="{% url 'clients:client_create' lang=lang %}"
                        type="button" style="max-width: 120px">
                         {% translate 'Create' %}
                     </a>
@@ -152,14 +152,14 @@
                                 <td>{{object.get_status_display}}</td>
                                 <td>
                                     <a class="btn btn-primary btn-sm btn-warning element-fullwidth"
-                                       href="{% url 'objects:selection' lang=lang client_id=object.id %}"
+                                       href="{% url 'clients:selection' lang=lang client_id=object.id %}"
                                        type="button" style="max-width: 120px">
                                         {% translate 'Selection' %}
                                     </a>
                                 </td>
                                 <td>
                                     <a class="btn btn-primary btn-sm btn-warning element-fullwidth"
-                                       href="{% url 'objects:selection_history' lang=lang pk=object.id %}"
+                                       href="{% url 'clients:selection_history' lang=lang pk=object.id %}"
                                        type="button" style="max-width: 120px">
                                         {% translate 'Selection history' %}
                                     </a>
@@ -167,14 +167,14 @@
                                 {% if can_update_clients and can_update_clients|get_dict_value:object.id %}
                                     <td>
                                         <a class="btn btn-primary btn-sm btn-warning element-fullwidth"
-                                           href="{% url 'handbooks:client_update' lang=lang pk=object.id %}"
+                                           href="{% url 'clients:client_update' lang=lang pk=object.id %}"
                                            type="button" style="max-width: 120px">
                                             {% translate 'Update' %}
                                         </a>
                                     </td>
                                     <td>
                                         <a class="btn btn-primary btn-sm btn-danger element-fullwidth"
-                                           href="{% url 'handbooks:client_delete' lang=lang pk=object.id %}"
+                                           href="{% url 'clients:client_delete' lang=lang pk=object.id %}"
                                            type="button" style="max-width: 120px">
                                             {% translate 'Delete' %}
                                         </a>
@@ -186,7 +186,7 @@
                                 {% if can_view_client_history %}
                                     <td>
                                         <a class="btn btn-primary btn-sm btn-warning element-fullwidth"
-                                           href="{% url 'handbooks:client_history' lang=lang pk=object.id %}"
+                                           href="{% url 'clients:client_history' lang=lang pk=object.id %}"
                                            type="button" style="max-width: 120px">
                                             {% translate 'History' %}
                                         </a>
