+++ clients/templates/clients/selection_history_list.html
@@ -9,7 +9,7 @@
     <h1>{% translate 'Selection history' %}</h1>
     <ul>
       <li><a href="{% url 'main' lang=lang %}">{% translate 'Main' %}</a></li>
-      <li><a href="{% url 'handbooks:all_client_list' lang=lang %}">{% translate 'Clients' %}</a></li>
+      <li><a href="{% url 'clients:all_client_list' lang=lang %}">{% translate 'Clients' %}</a></li>
       <li>/ {% translate 'Selection history' %}</li>
     </ul>
   </div>
