+++ objects/admin.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
 
-from objects.models import Apartment, Commerce, House, Selection
+from . import models
 
-admin.site.register(Apartment)
-admin.site.register(Commerce)
-admin.site.register(House)
-admin.site.register(Selection)
+admin.site.register(models.Apartment)
+admin.site.register(models.Commerce)
+admin.site.register(models.House)
+admin.site.register(models.Land)
