+++ objects/models.py
@@ -5,18 +5,25 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from simple_history.models import HistoricalRecords
 
-from accounts.models import CustomUser
-from handbooks.models import Client, Handbook, Locality, Street
 from images.models import RealEstateImage
-
-from .choices import RealEstateStatus, LandTarget, LandDisposition, LandRubric, HouseRubric, CommerceRubric, \
-    ApartmentRubric, RealEstateDocument, RealEstateCommunication, HouseRoomsNumberRubric
+from .choices import (
+    RealEstateStatus,
+    LandTarget,
+    LandDisposition,
+    LandRubric,
+    HouseRubric,
+    CommerceRubric,
+    ApartmentRubric,
+    RealEstateDocument,
+    RealEstateCommunication,
+    HouseRoomsNumberRubric
+)
 
 
 class BaseRealEstate(models.Model):
     """
     Базовий клас, який містить спільні поля для
-    обʼєктів нерухомості: квартири, комерції та будинку.
+    обʼєктів нерухомості: квартири, комерції, будинку та земельної ділянки.
     """
 
     class Meta:
@@ -25,7 +32,7 @@ class BaseRealEstate(models.Model):
 
     creation_date = models.DateField(
         verbose_name=_("Creation date"), default=datetime.date.today
-    )  # дата cтворення
+    )  # дата створення
     deposit_date = models.DateField(
         null=True, blank=True, verbose_name=_("Deposit date")
     )  # дата постановки
@@ -50,17 +57,17 @@ class BaseRealEstate(models.Model):
     #     null=True, blank=True,
     # )
     locality = models.ForeignKey(
-        Locality, on_delete=models.CASCADE, verbose_name=_("Locality")
+        "handbooks.Locality", on_delete=models.CASCADE, verbose_name=_("Locality")
     )
     # locality_district = models.ForeignKey(
     #     LocalityDistrict,
     #     on_delete=models.CASCADE,
     # )
-    street = models.ForeignKey(Street, on_delete=models.CASCADE, verbose_name=_("Street"))
+    street = models.ForeignKey("handbooks.Street", on_delete=models.CASCADE, verbose_name=_("Street"))
     house = models.CharField(max_length=100, verbose_name=_("House"))
 
     realtor = models.ForeignKey(
-        CustomUser,
+        "accounts.CustomUser",
         on_delete=models.CASCADE,
         related_name="%(app_label)s_%(class)ss",
         verbose_name=_("Realtor"),
@@ -78,7 +85,7 @@ class BaseRealEstate(models.Model):
     #     related_name="%(app_label)s_%(class)ss",
     # )
     condition = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="condition_%(app_label)s_%(class)ss",
         verbose_name=_("Condition"),
@@ -86,7 +93,7 @@ class BaseRealEstate(models.Model):
         blank=True,
     )
     material = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="material_%(app_label)s_%(class)ss",
         verbose_name=_("Material"),
@@ -94,7 +101,7 @@ class BaseRealEstate(models.Model):
         blank=True,
     )
     agency = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="agency_%(app_label)s_%(class)ss",
         verbose_name=_("Agency"),
@@ -106,7 +113,7 @@ class BaseRealEstate(models.Model):
         verbose_name=_("Filial agency"),
     )
     house_type = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="house_type_%(app_label)s_%(class)ss",
         verbose_name=_("House type"),
@@ -114,7 +121,7 @@ class BaseRealEstate(models.Model):
         blank=True,
     )
     layout = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="layout_%(app_label)s_%(class)ss",
         verbose_name=_("Layout"),
@@ -122,7 +129,7 @@ class BaseRealEstate(models.Model):
         blank=True,
     )
     stair = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="stair_%(app_label)s_%(class)ss",
         verbose_name=_("Stair"),
@@ -143,7 +150,7 @@ class BaseRealEstate(models.Model):
     # )
 
     owner = models.ForeignKey(
-        Client,
+        "clients.Client",
         on_delete=models.CASCADE,
         related_name="owner_%(app_label)s_%(class)ss",
         verbose_name=_("Owner"),
@@ -254,7 +261,7 @@ class Apartment(BaseRealEstate):
         default=0, verbose_name=_("Number of balconies")
     )
     complex = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="complex_objects_apartment",
         verbose_name=_("Complex"),
@@ -323,7 +330,7 @@ class Commerce(BaseRealEstate):
     )
     # office = models.BooleanField(default=False) # квартира під офіс
     complex = models.ForeignKey(
-        Handbook,
+        "handbooks.Handbook",
         on_delete=models.CASCADE,
         related_name="complex_objects_commerce",
         verbose_name=_("Complex"),
@@ -389,24 +396,3 @@ class Land(BaseRealEstate):
         choices=LandDisposition.choices, verbose_name=_("Disposition")
     )  # розташування
     own_parking = models.BooleanField(default=False, verbose_name=_("Own parking"))
-
-
-class Selection(models.Model):
-    class Meta(BaseRealEstate.Meta):
-        permissions = (("selection", "Selection"),)
-
-    client = models.ForeignKey(Client, on_delete=models.CASCADE, verbose_name=_("Client"))
-    date = models.DateField(default=datetime.date.today, verbose_name=_("Date"))
-    user = models.ForeignKey(CustomUser, on_delete=models.CASCADE, verbose_name=_("User"))
-    selected_apartments = models.ManyToManyField(
-        Apartment, blank=True, related_name="related_selected_apartments"
-    )
-    selected_houses = models.ManyToManyField(
-        House, blank=True, related_name="related_selected_houses"
-    )
-    selected_commerces = models.ManyToManyField(
-        Commerce, blank=True, related_name="related_selected_commerces"
-    )
-    selected_lands = models.ManyToManyField(
-        Land, blank=True, related_name="related_selected_lands"
-    )
