+++ templates/main.html
@@ -26,7 +26,7 @@
                 <div class="featured_content">
                     <h2>{% translate 'Selections list' %}</h2>
                     <div class="explore_btn">
-                        <a class="custombtn" href="{% url 'handbooks:all_client_list' lang=lang %}">
+                        <a class="custombtn" href="{% url 'clients:all_client_list' lang=lang %}">
                             {% translate 'View' %}
                         </a>
                     </div>
